@@ -1,10 +1,11 @@
 {
   "name": "npm-package-downloader",
-  "version": "0.1.6",
+  "version": "0.2.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=14.0.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -25,11 +26,12 @@
   "dependencies": {
     "bluebird": "^2.10.2",
     "commander": "^2.9.0",
-    "download": "^4.3.1",
+    "download": "^4.4.3",
     "download-status": "^2.2.1",
     "got": "^5.0.0",
-    "lodash": "^3.10.1",
+    "lodash": "^4.17.21",
     "node-7z": "^0.3.0",
+    "registry-url": "^6.0.0",
     "semver": "^5.0.3",
     "valid-url": "^1.0.9",
     "winston": "^1.1.1"
