@@ -1,9 +1,9 @@
 "use strict";
-let winston = require('winston'),
-    Zip = require('node-7z');
-const path = require('path');
+import winston from "winston";
+import Zip from "node-7z";
+import path from "path";
 
-module.exports = function (location) {
+export default function (location) {
     winston.info('7Ziping your files, Please wait...');
 
     let zip = new Zip();
